chore: SLOP cleanup, deduplicate tool utilities, bump to v5.7.1
Extract safe_read, max_file_size, sensitive_file? to BaseTool (16 duplicate
one-liners removed across 8 tools). Extract FullSerializerBehavior module
for shared footer/architecture_summary. Derive tools_name_list from TOOL_ROWS.
Fix notable_gems_list bypass in 2 serializers. Narrow bare rescue to
rescue StandardError (4 sites). Delete dead SENSITIVE_PATTERNS constant.

20 files changed, +79/-155 lines (net -76).
