+++ lib/rails_ai_context/serializers/claude_serializer.rb
@@ -104,6 +104,8 @@ module RailsAiContext
 
     # Internal: full-mode Claude serializer (wraps MarkdownSerializer with behavioral rules)
     class FullClaudeSerializer < MarkdownSerializer
+      include FullSerializerBehavior
+
       private
 
       def header
@@ -118,27 +120,6 @@ module RailsAiContext
           structure, patterns, and conventions. Use it to write idiomatic code
           that matches this project's style.
         MD
-      end
-
-      def footer
-        rules = []
-        rules << "## Behavioral Rules"
-        rules << ""
-        rules << "When working in this codebase:"
-        rules << "- Follow existing patterns and conventions detected above"
-        rules << "- Use the database schema as the source of truth for column names and types"
-        rules << "- Respect existing associations and validations when modifying models"
-        rules << "- Match the project's architecture style (#{architecture_summary})" if architecture_summary
-        test_cmd = detect_test_command
-        rules << "- Run `#{test_cmd}` after making changes to verify correctness"
-        rules << ""
-        rules << super
-        rules.join("\n")
-      end
-
-      def architecture_summary
-        arch = context.dig(:conventions, :architecture)
-        arch&.any? ? arch.join(", ") : nil
       end
     end
   end
