+++ lib/rails_ai_context/tools/get_job_pattern.rb
@@ -404,14 +404,6 @@ module RailsAiContext
 
         enqueuers.to_a.sort.first(20)
       end
-
-      private_class_method def self.safe_read(path)
-        RailsAiContext::SafeFile.read(path)
-      end
-
-      private_class_method def self.max_file_size
-        RailsAiContext.configuration.max_file_size
-      end
     end
   end
 end
