+++ lib/rails_ai_context/tools/get_partial_interface.rb
@@ -470,14 +470,6 @@ module RailsAiContext
         $stderr.puts "[rails-ai-context] find_available_partials failed: #{e.message}" if ENV["DEBUG"]
         []
       end
-
-      private_class_method def self.safe_read(path)
-        RailsAiContext::SafeFile.read(path)
-      end
-
-      private_class_method def self.max_file_size
-        RailsAiContext.configuration.max_file_size
-      end
     end
   end
 end
