+++ lib/rails_ai_context/tools/get_service_pattern.rb
@@ -316,14 +316,6 @@ module RailsAiContext
 
         parts.any? ? parts.join(", ") : "mixed/no dominant pattern"
       end
-
-      private_class_method def self.safe_read(path)
-        RailsAiContext::SafeFile.read(path)
-      end
-
-      private_class_method def self.max_file_size
-        RailsAiContext.configuration.max_file_size
-      end
     end
   end
 end
